Refactor data to include all game info, send to analysis page
The state keeping track of each game is more complex, now containing 63 arrays
instead of 63 winners. Each array contains the winner, as well as the two teams playing.
This allows the results page to tell if/how big of an upset a particular game was.
Pass this data to analysis page to use.